build: add packaging support and rename build script
closes: #12

This patch updates the build files to include preliminary build and
installation support.  This patch also renames the build script from
'build-deps.sh' to 'build.sh' to better define it's actual role in
the build process.  This patch also updates the travis.yml automated
build to use this new script and updates the readme file with new
instructions.

Signed-off-by: Keith Holman <keith.holman@windriver.com>