tr50: remove extra "\n" from log messages
Some logs in the tr50 plug-in included a trailing new-line (\n)
character.  One is automatically appended when outputting the log
message, so this causes blank lines to be outputted in the log.  This
fix removes these trailing "\n".  Additionally, 1 os_printf statement
was updated to b a trace log message to allow it to be turned on or off.

Signed-off-by: Keith Holman <keith.holman@windriver.com>