build: remove octals numbers splitting version
The algorithm used to split version numbers was simply diving at the "."
character.  A new fix introduced generated the project version from the
date, leading to a version of: "17.12.08".  The 08, was then being split
into a number 08, which to C code would be the representation for an
invalid octal value (of 8).  This fix updates the split algorithm to
ignore leading 0's from version splitting, starting a the first number
value with a 1-9.

Signed-off-by: Keith Holman <keith.holman@windriver.com>