control: use api to retrieving correct directories
This patch updates the code to retrieve the correct directories where
configuration file and runtime files are stored.  Originally, this code
was working by using the default directories.  However, on some systems
these directories might be changed.  This change updates the source code
to use an API call to retrieve these directories.

Signed-off-by: Keith Holman <keith.holman@windriver.com>