api: handle case when no space in wait queue
closes: #63

This patch updates the action code within the api to handle the case
where there is no space on the request wait queue.  Previously, if this
happened the queue would roll around replacing the oldest request on the
queue even if it was currently in progress.  This fix will allow the api
to return an error indicating the queue is full to the cloud.

Signed-off-by: Keith Holman <keith.holman@windriver.com>