test: update alarm unit tests
This patch updates the alarm unit tests to provide 100% coverage, it
includes minor fixes to the code base.  There are fixes around
preventing memory leaks, in rare cases where the system is nearly
exhausted of memory.  There is also a fix that reduces the requirement
to allocate memory on the heap in order to send an alarm.

Signed-off-by: Keith Holman <keith.holman@windriver.com>