tr50: executing non-cloud based actions causes segfault
closes: #57

The tr50 plug-in code worked assuming all actions were initiated by the
cloud.  This is not the case, by call iot_action_execute directly one is
able to similate and action locally.  However, doing this was causing a
seg fault as the action was trying to update the cloud with a result.
This patch provides a fix that disables this update as intended.

Signed-off-by: Keith Holman <keith.holman@windriver.com>