build: correct version split for: 10, 20 & 30
The splitting of a version string to into their equivilent numeric portions
was incorrectly parsing numbers by removing all zeros.  For example the
version For example, the version "18.01.30" was being parsed into the numerals:
18, 1, 3 & 0, respecfully.  However, the desired solution should be: 18,
1, 30 & 0.  This patch updates the script to handle the case for the numbers
in this situation.  Since version is based on a calendar date, this set
only includes the numbers: 10, 20 & 30.

Signed-off-by: Keith Holman <keith.holman@windriver.com>