device-manager: publish api version to api_version
Previously the API version was being published to an attribute named:
`device-cloud_version`, however, this was not in the thing definition.
The thing definition contained an item named: `hdc_version`, but the
project is in the process of removing the `hdc` name.  This patch
updates the attribute to be published as the property `api_version` so
that the property is now name agnostic.

Signed-off-by: Keith Holman <keith.holman@windriver.com>