device-manager: add new default action Ping Device.
Add new default action "Ping Device".  This action is intended to test
the transport from cloud to agent and return a timestamp in UTC.  This
action is suitable to be called from a trigger for testing uptime on
devices.  This action is also a diagnostic utility for testing
action sanity on the device manager.
Fixed a typo in a struct name.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>