api: remove internal connection changed tracking
The mqtt portion of the library was tracking a variable called
`connection_changed` this variable was being set to update when the
connection status of a client changed.  However, this was only relevent
for 1 iteration, and the mqtt client has no way of knowing when an
iteration change happens.  Thus, this state tracking was causing
internal issues of when to call a reconnect.  This information should be
tracked outside of the mqtt portion of the library to be implemented
correctly.

Signed-off-by: Keith Holman <keith.holman@windriver.com>