relay: allow the use of civetweb as a websocket library
This patch updates the relay source code to allow the use of civetweb as
a websocket client library.  This is desired as civetweb has a more
liberal license (MIT) versus libwebsockets (GPL) license that is
required by some clients.  However, there are limitations of using
civetweb such as using SSL certificates as civetweb is designed to be an
embedded webserver, and not a websocket client library.

Signed-off-by: Keith Holman <keith.holman@windriver.com>