device-manager: incorporate some vxworks changes
This patch updates the device manager, as well as the complete app with
some of the changes to work on VxWorks.  The purpose of this is to make
the delta between the VxWorks layer and the master branch smaller to
facilitate easier porting of the code base to a VxWorks based system.

Signed-off-by: Keith Holman <keith.holman@windriver.com>