api: Add mutex lock for action completion in plugins
This adds a mutex lock around the performing of the action complete operation
of a plugin, as part of the mailbox protocol changes.

Previously it was possible for a race condition to occur when
two actions completed at exactly the same time, causing an invalid
double check.

This patch also changes the mailbox check on received method requests
to only occur after the received request is already allocated. This change,
along with the lock around the action complete operation should prevent
the "maximum inbound requests reached" message from appearing.