relay: fix segmentation fault if no url specified
This patch updates the iot-relay application to not produce a
segmentation fault if not remote access url is provided.  The
segmentation fault is caused by the argument parser returns the index
where the first non-parameterized argument should be.  However, the code
isn't checking if this argument is actually specified a value and
assumes it contains the web address.  Then connecting with some
underlying websocket libraries is causing a segmentation fault.  The fix
is to correctly check for a address value of NULL in the parameter line.

Signed-off-by: Keith Holman <keith.holman@windriver.com>