build: use same branch name on OSAL if exists
There have been times where signatures to functions have changed on the
OSAL layer, but that breaks the hdc-lib code base until changes are made
on that repository.  This patch updates the build script to use change
to branch of the same name on the OSAL layer, if it exists.  Using this
method if a signature change is being performed on the OSAL layer the
change can be done on the HDC-repo in a branch of the same time to test
both repository prior to submitting the code.   When both pull requests
pass review then they can be pushed nearly the same time to lower the
change of having a broken cloud repository.

Signed-off-by: Keith Holman <keith.holman@windriver.com>