relay: pass ca certificate bundle as argument
This patch updates the relay application to pass the Certicate Authroity
(CA) bundle as an argument on the command line.  This ensures that the
device manager and relay are using the same authentication mechanisms.
Previously, the relay app was trying to read the configuration file
directly.  However, if the device manager was configured with it's own
configuration file "overriding" the default, then the iot-relay was not
being passed this information and would not be able to connect.  This
patch allivates this problem.

Signed-off-by: Keith Holman <keith.holman@windriver.com>