device manager: add action to update remote login
This patch updates the method for reading the supported remote login
protocols from the `iot.cfg` configuration file.  They are now added
into an action to align with how the python agent works.  This will
allows the cloud the be able to have the device update the supported
list of protocols via an action that will keep the list in remote access
up-to-date.

Signed-off-by: Keith Holman <keith.holman@windriver.com>